暂存，可能改的有点多，先暂存，和存储逻辑有关

+++ TripleBit/HashIndex.cpp
@@ -336,13 +336,13 @@ unsigned HashIndex::next(ID id)
 	return hashTableEntries[firstHash];
 }
 
-char* writeData(char* writer, unsigned long long data)
+char* writeData(char* writer, size_t data)
 {
 	memcpy(writer, &data, sizeof(data));
 	return writer+ sizeof(data);
 }
 
-const char* readData(const char* reader, unsigned long long& data)
+const char* readData(const char* reader, size_t& data)
 {
 	memcpy(&data, reader, sizeof(data));
 	return reader+ sizeof(data);
